Allow property access from traits
